Increase clientTick to UInt32, fix UInt16 rollover, and up rolling average window of latency calculations to 15 instead of 3

+++ src/common/schemas/clientTickSchema.ts
@@ -2,5 +2,5 @@ import { defineSchema } from '../binary/schema/defineSchema'
 import { Binary } from '../binary/Binary'
 
 export const clientTickSchema = defineSchema({
-    tick: Binary.UInt16,
+    tick: Binary.UInt32,
 })
